Almost Done Backend

+++ LandAdministration/src/main/java/com/example/landadministration/services/LandService.java
@@ -1,16 +1,31 @@
 package com.example.landadministration.services;
 
 import com.example.landadministration.dtos.LandDTO;
-import com.example.landadministration.entities.Land;
+import com.example.landadministration.dtos.LandOwnerDTO;
+import com.example.landadministration.entities.*;
+import com.example.landadministration.repos.LandOwnerRepo;
 import com.example.landadministration.repos.LandRepo;
+import com.example.landadministration.repos.OwnershipHistoryRepo;
+import com.example.landadministration.repos.UserLogRepo;
+import com.example.landadministration.specifications.LandSpecification;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import net.minidev.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.web.PagedModel;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.Period;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -18,8 +33,22 @@ import java.util.Optional;
 @Service
 public class LandService {
 
+    @Value("${opencage.api.key}")
+    private String apiKey;
+
     @Autowired
     private LandRepo landRepo;
+
+    @Autowired
+    private LandOwnerRepo landOwnerRepo;
+
+    @Autowired
+    private OwnershipHistoryRepo ownershipHistoryRepo;
+
+    @Autowired
+    private UserLogRepo userLogRepo;
+
+
 
 
     public List<LandDTO> getLandRecords(String sortedBy) {
@@ -37,9 +66,6 @@ public class LandService {
     }
 
     public LandDTO addRecord(Land land) {
-        if(land.getLocation()==null){
-            throw new IllegalStateException("Location not entered");
-        }
         if (Double.isNaN(land.getLongitude())) {
             throw new IllegalStateException("Longitude not entered");
         }
@@ -55,9 +81,50 @@ public class LandService {
         if (land.getUsage_type() == null) {
             land.setUsage_type("Residential");
         }
+        land.setLocation(getLocationFromCoordinates(land.getLatitude(), land.getLongitude()));
+        Users userNavigating = (Users) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        UserLog userLog = new UserLog();
+        userLog.setUser(userNavigating);
+        userLog.setAction("ADD_LAND");
+        userLog.setTimestamp(LocalDateTime.now());
+        userLog.setDescription("User {" + userNavigating.getUsername() + "} added land with id " + land.getId() + ".");
+        userLogRepo.save(userLog);
         Land savedLand = landRepo.save(land);
         return getDTO(savedLand);
     }
+
+    public String getLocationFromCoordinates(double latitude, double longitude) {
+        try {
+            String url = String.format(
+                    "https://api.opencagedata.com/geocode/v1/json?q=%f,%f&key=%s",
+                    latitude, longitude, apiKey
+            );
+
+            RestTemplate restTemplate = new RestTemplate();
+            String response = restTemplate.getForObject(url, String.class);
+
+            ObjectMapper mapper = new ObjectMapper();
+            JsonNode root = mapper.readTree(response);
+
+            JsonNode components = root
+                    .path("results").get(0)
+                    .path("components");
+
+            String city = components.path("village").asText(null);
+            if (city == null) city = components.path("town").asText(null);
+            if (city == null) city = components.path("city").asText("Unknown");
+
+            String state = components.path("state").asText("");
+            String country = components.path("country").asText("");
+
+            return String.join(", ", city, state, country).replaceAll("(^,\\s*)|(,\\s*$)", "");
+
+        } catch (Exception e) {
+            System.err.println("Reverse geocoding failed: " + e.getMessage());
+            return "Unknown";
+        }
+    }
+
 
     public LandDTO updateUsageType(Integer id, String usageType) {
         Optional<Land> landOptional = landRepo.findById(id);
@@ -66,8 +133,16 @@ public class LandService {
         }
         Land landToUpdate = landOptional.get();
         if(landToUpdate.getUsage_type() != null){
+            Users userNavigating = (Users) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+            UserLog userLog = new UserLog();
+            userLog.setUser(userNavigating);
+            userLog.setAction("UPDATE_USAGE_TYPE");
+            userLog.setTimestamp(LocalDateTime.now());
+            userLog.setDescription("User {" + userNavigating.getUsername() + "} updated usage type of land with id " +
+                    id + "from {" + landToUpdate.getUsage_type() +"} to {" + usageType + "}.");
             landToUpdate.setUsage_type(usageType);
             Land updatedLand = landRepo.save(landToUpdate);
+            userLogRepo.save(userLog);
             return getDTO(updatedLand);
         }
         throw new IllegalStateException("Usage type not entered");
@@ -87,7 +162,17 @@ public class LandService {
         if(!landOptional.isPresent()){
             throw new IllegalStateException("Land not found");
         }
+        if(landOptional.get().getLandOwner()!=null){
+            throw new IllegalStateException("Land owner is assigned to this land, cannot delete the land. Please unassign the land owner first and then delete the land.");
+        }
+        Users userNavigating = (Users) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        UserLog userLog = new UserLog();
+        userLog.setUser(userNavigating);
+        userLog.setAction("DELETE_LAND");
+        userLog.setTimestamp(LocalDateTime.now());
+        userLog.setDescription("User {" + userNavigating.getUsername() + "} deleted land with id {" + id + "}.");
         landRepo.deleteById(id);
+        userLogRepo.save(userLog);
         Land landToDelete = landOptional.get();
         return getDTO(landToDelete);
     }
@@ -166,7 +251,8 @@ public class LandService {
                     land.getId(),
                     land.getLocation(),
                     land.getSurfaceArea(),
-                    land.getUsage_type()
+                    land.getUsage_type(),
+                    getDTO(land.getLandOwner())
             );
             dto.setLocationCoordinates(land.getLatitude(), land.getLongitude());
             return dto;
@@ -194,7 +280,7 @@ public class LandService {
         }
 
         return landPage.map(land -> {
-            LandDTO dto = new LandDTO(land.getId(), land.getLocation(), land.getSurfaceArea(), land.getUsage_type());
+            LandDTO dto = new LandDTO(land.getId(), land.getLocation(), land.getSurfaceArea(), land.getUsage_type(),getDTO(land.getLandOwner()));
             dto.setLocationCoordinates(land.getLatitude(), land.getLongitude());
             return dto;
         });
@@ -225,7 +311,7 @@ public class LandService {
         }
 
         return landPage.map(land -> {
-            LandDTO dto = new LandDTO(land.getId(), land.getLocation(), land.getSurfaceArea(), land.getUsage_type());
+            LandDTO dto = new LandDTO(land.getId(), land.getLocation(), land.getSurfaceArea(), land.getUsage_type(),getDTO(land.getLandOwner()));
             dto.setLocationCoordinates(land.getLatitude(), land.getLongitude());
             return dto;
         });
@@ -254,7 +340,7 @@ public class LandService {
         }
 
         return landPage.map(land -> {
-            LandDTO dto = new LandDTO(land.getId(), land.getLocation(), land.getSurfaceArea(), land.getUsage_type());
+            LandDTO dto = new LandDTO(land.getId(), land.getLocation(), land.getSurfaceArea(), land.getUsage_type(),getDTO(land.getLandOwner()));
             dto.setLocationCoordinates(land.getLatitude(), land.getLongitude());
             return dto;
         });
@@ -263,16 +349,78 @@ public class LandService {
     public List<LandDTO> getDTOList(List<Land> lands){
         List<LandDTO> landDTOList = new ArrayList<>();
         for(Land land : lands){
-            LandDTO landDTO = new LandDTO(land.getId(),land.getLocation(),land.getSurfaceArea(),land.getUsage_type());
+            LandDTO landDTO = new LandDTO(land.getId(),land.getLocation(),land.getSurfaceArea(),land.getUsage_type(),getDTO(land.getLandOwner()));
             landDTO.setLocationCoordinates(land.getLatitude(),land.getLongitude());
             landDTOList.add(landDTO);
         }
         return landDTOList;
     }
 
+
     public LandDTO getDTO(Land land){
-        LandDTO landDTO = new LandDTO(land.getId(),land.getLocation(),land.getSurfaceArea(),land.getUsage_type());
+        LandDTO landDTO = new LandDTO(land.getId(),land.getLocation(),land.getSurfaceArea(),land.getUsage_type(),getDTO(land.getLandOwner()));
         landDTO.setLocationCoordinates(land.getLatitude(),land.getLongitude());
         return landDTO;
+    }
+    public Page<LandDTO> searchLands(String location, String usageType, String ownerName, String sortedBy, int page, int size) {
+        Specification<Land> spec = (root, query, cb) -> cb.conjunction();
+
+        if (location != null && !location.isEmpty()) {
+            spec = spec.and(LandSpecification.hasLocation(location));
+        }
+        if (usageType != null && !usageType.isEmpty()) {
+            spec = spec.and(LandSpecification.hasUsageType(usageType));
+        }
+        if (ownerName != null && !ownerName.isEmpty()) {
+            spec = spec.and(LandSpecification.hasOwnerName(ownerName));
+        }
+
+        Sort sort = Sort.by(Sort.Direction.ASC, sortedBy);
+        Pageable pageable = PageRequest.of(page, size, sort);
+        Page<Land> landPage = landRepo.findAll(spec, pageable);
+
+        return landPage.map(land -> {
+            LandDTO dto = new LandDTO(land.getId(), land.getLocation(), land.getSurfaceArea(), land.getUsage_type(),getDTO(land.getLandOwner()));
+            dto.setLocationCoordinates(land.getLatitude(), land.getLongitude());
+            return dto;
+        });
+    }
+    public LandOwnerDTO getDTO(LandOwner landOwner){
+        if(landOwner==null){
+            return null;
+        }
+        Integer landCount = 0;
+        if(landOwner.getLands()!=null){
+            landCount = landOwner.getLands().size();
+        }
+        LandOwnerDTO landOwnerDTO =  new LandOwnerDTO(landOwner.getId(),
+                landOwner.getFirstName()+" "+landOwner.getLastName(),
+                landOwner.getPhoneNb(),
+                landOwner.getEmailAddress(),
+                landCount,
+                Period.between(landOwner.getDateOfBirth(), LocalDate.now()).getYears());
+        return landOwnerDTO;
+    }
+
+    public LandDTO unassignOwner(Integer landId) {
+        Land land = landRepo.findById(landId)
+                .orElseThrow(() -> new IllegalStateException("Land not found"));
+        if(land.getLandOwner()==null){
+            throw new IllegalStateException("There is no land owner assigned to this land");
+        }
+        LandOwner owner = land.getLandOwner();
+
+        OwnershipHistoryId id = new OwnershipHistoryId(land.getId(), owner.getId());
+
+        Optional<OwnershipHistory> historyOptional = ownershipHistoryRepo.findById(id);
+        historyOptional.get().setOwnershipEnd(LocalDateTime.now());
+        land.setLandOwner(null);
+        Land updatedLand = landRepo.save(land);
+        return getDTO(updatedLand);
+    }
+
+    public Integer getUnassignedLands() {
+        System.out.println("Unassigned lands: " + landRepo.getUnassignedLands().size());
+        return landRepo.getUnassignedLands().size();
     }
 }
